requirements.txt
selenium
beautifulsoup4
pandas
openai
python-dotenv
webdriver-manager 